requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
statsmodels>=0.13.0
prophet>=1.0.1
python-dotenv>=0.19.0
pydantic>=1.8.2
pytest>=6.2.5
httpx>=0.18.2
pandas-ta>=0.3.14b
plotly>=5.3.1
python-multipart>=0.0.5
joblib>=1.1.0
redis>=4.0.0
celery>=5.1.2
flower>=1.0.0
